--- src/components/CommonModal/CommonModal.css
@@ -1,30 +1,0 @@
-.companyList {
-  border: 1px solid #bcbcbc;
-  border-radius: 5px;
-  margin-right: 20px;
-  height: 510px;
-  width: 38%;
-  margin-left: 10px;
-  height: 580px;
-}
-
-.checkbox {
-  border-radius: 5px;
-  height: 200px;
-  margin-bottom: 25px;
-}
-
-.choiceEmp {
-  border: 1px solid #bcbcbc;
-  border-radius: 5px;
-  height: 320px;
-  margin-top: 10px;
-}
-
-container {
-  padding: 2rem 0rem;
-}
-
-h4 {
-  margin: 2rem 0rem 1rem;
-}
