--- src/components/Modals/MyGroup.js
@@ -1,5 +1,0 @@
-const MyGroup = () => {
-  return <div>mygroup page test</div>;
-};
-
-export default MyGroup;
